web/requirements.txt
django
lxml
django-ninja
cssselect
whitenoise
daphne
